move version error into a dedicated module
