+++ types/src/errors/qs.rs
@@ -5,7 +5,7 @@
 use thiserror::Error;
 use tls_codec::{TlsDeserializeBytes, TlsSerialize, TlsSize};
 
-use crate::messages::client_qs::VersionError;
+use super::version::VersionError;
 
 /// Error fetching a message from the QS.
 #[derive(Error, Debug, Clone, TlsSerialize, TlsDeserializeBytes, TlsSize)]
